requirements.txt
python-telegram-bot==21.6
yfinance==0.2.40
matplotlib==3.9.2
numpy==2.1.1
pandas==2.2.3
